Code quality improvements
Use of 1:nrow(x) is not a good practice, as in the case when
"nrow(x) == 0" the for loop will run for "1:0", and that is two
iterations (i==1, i==0). Usually the developer intended that the for
loop run 0 times if x had 0 rows. using "seq_len(nrow(x))" is the
recommended practice to generate an integer based sequence starting
at one and with the length equal to the number of rows of x.

"any(is.na(x))"  can be replaced by "anyNA(x)" since R 3.2. The replacement
should be faster.