Add continuous integration support
Continuous Integration (CI) means that commits on this repository will automatically be tested.

travis-ci.org is a company that provides CI support for free for open source projects, with good defaults for multiple programming languages.

By default, Travis builds and checks the R package (`language: r` tells Travis this is an R repository)

Besides, we also add a code coverage report, that tells which lines of code are run when we check the package.

And a pkgdown website on github pages

apt dependencies are used to install dependencies faster