Add `ppm_VIP_vector` function
The function extracts the VIPs vector (numeric) from the model_VIP(MVObj) function. It is not necessary if you have the ppm values in a numeric vector. This is needed in case that an automated pipeline is applied, connecting the output from model_VIP(MVObj) to nmr_identify_regions family functions.